+++ modules/planning/common/path/discretized_path.cc
@@ -82,12 +82,12 @@ const std::vector<common::PathPoint> &DiscretizedPath::path_points() const {
   return path_points_;
 }
 
-std::size_t DiscretizedPath::num_of_points() const {
+std::uint32_t DiscretizedPath::num_of_points() const {
   return path_points_.size();
 }
 
 const common::PathPoint &DiscretizedPath::path_point_at(
-    const std::size_t index) const {
+    const std::uint32_t index) const {
   CHECK_LT(index, path_points_.size());
   return path_points_[index];
 }
@@ -102,7 +102,7 @@ common::PathPoint DiscretizedPath::end_point() const {
   return path_points_.back();
 }
 
-common::PathPoint &DiscretizedPath::path_point_at(const std::size_t index) {
+common::PathPoint &DiscretizedPath::path_point_at(const std::uint32_t index) {
   CHECK_LT(index, path_points_.size());
   return path_points_[index];
 }
