+++ modules/planning/math/smoothing_spline/spline_2d.cc
@@ -26,10 +26,11 @@
 namespace apollo {
 namespace planning {
 
-Spline2d::Spline2d(const std::vector<double>& t_knots, const std::size_t order)
+Spline2d::Spline2d(const std::vector<double>& t_knots,
+                   const std::uint32_t order)
     : t_knots_(t_knots), spline_order_(order) {
   if (t_knots.size() > 1) {
-    for (std::size_t i = 1; i < t_knots_.size(); ++i) {
+    for (std::uint32_t i = 1; i < t_knots_.size(); ++i) {
       splines_.emplace_back(spline_order_);
     }
   }
@@ -39,7 +40,7 @@ std::pair<double, double> Spline2d::operator()(const double t) const {
   if (splines_.size() == 0) {
     return std::make_pair(0.0, 0.0);
   }
-  std::size_t index = find_index(t);
+  std::uint32_t index = find_index(t);
   return splines_[index](t - t_knots_[index]);
 }
 
@@ -47,7 +48,7 @@ double Spline2d::x(const double t) const {
   if (splines_.size() == 0) {
     return 0.0;
   }
-  std::size_t index = find_index(t);
+  std::uint32_t index = find_index(t);
   return splines_[index].x(t - t_knots_[index]);
 }
 
@@ -55,7 +56,7 @@ double Spline2d::y(const double t) const {
   if (splines_.size() == 0) {
     return 0.0;
   }
-  std::size_t index = find_index(t);
+  std::uint32_t index = find_index(t);
   return splines_[index].y(t - t_knots_[index]);
 }
 
@@ -64,7 +65,7 @@ double Spline2d::derivative_x(const double t) const {
   if (splines_.size() == 0) {
     return 0.0;
   }
-  std::size_t index = find_index(t);
+  std::uint32_t index = find_index(t);
   return splines_[index].derivative_x(t - t_knots_[index]);
 }
 
@@ -73,7 +74,7 @@ double Spline2d::derivative_y(const double t) const {
   if (splines_.size() == 0) {
     return 0.0;
   }
-  std::size_t index = find_index(t);
+  std::uint32_t index = find_index(t);
   return splines_[index].derivative_y(t - t_knots_[index]);
 }
 
@@ -81,7 +82,7 @@ double Spline2d::second_derivative_x(const double t) const {
   if (splines_.size() == 0) {
     return 0.0;
   }
-  std::size_t index = find_index(t);
+  std::uint32_t index = find_index(t);
   return splines_[index].second_derivative_x(t - t_knots_[index]);
 }
 
@@ -89,7 +90,7 @@ double Spline2d::second_derivative_y(const double t) const {
   if (splines_.size() == 0) {
     return 0.0;
   }
-  std::size_t index = find_index(t);
+  std::uint32_t index = find_index(t);
   return splines_[index].second_derivative_y(t - t_knots_[index]);
 }
 
@@ -97,7 +98,7 @@ double Spline2d::third_derivative_x(const double t) const {
   if (splines_.size() == 0) {
     return 0.0;
   }
-  std::size_t index = find_index(t);
+  std::uint32_t index = find_index(t);
   return splines_[index].third_derivative_x(t - t_knots_[index]);
 }
 
@@ -105,23 +106,23 @@ double Spline2d::third_derivative_y(const double t) const {
   if (splines_.size() == 0) {
     return 0.0;
   }
-  std::size_t index = find_index(t);
+  std::uint32_t index = find_index(t);
   return splines_[index].third_derivative_y(t - t_knots_[index]);
 }
 /**
 *   @brief: set splines
 **/
 bool Spline2d::set_splines(const Eigen::MatrixXd& params,
-                           const std::size_t order) {
+                           const std::uint32_t order) {
   // check if the parameter size fit
   if (2 * t_knots_.size() * order !=
-      2 * order + static_cast<std::size_t>(params.rows())) {
+      2 * order + static_cast<std::uint32_t>(params.rows())) {
     return false;
   }
-  for (std::size_t i = 0; i < splines_.size(); ++i) {
+  for (std::uint32_t i = 0; i < splines_.size(); ++i) {
     std::vector<double> spline_piece_x(order, 0.0);
     std::vector<double> spline_piece_y(order, 0.0);
-    for (std::size_t j = 0; j < order; ++j) {
+    for (std::uint32_t j = 0; j < order; ++j) {
       spline_piece_x[j] = params(2 * i * order + j, 0);
       spline_piece_y[j] = params((2 * i + 1) * order + j, 0);
     }
@@ -133,22 +134,23 @@ bool Spline2d::set_splines(const Eigen::MatrixXd& params,
 
 // get the mutable single smoothing spline, if index out of range, nullptr will
 // returned;
-Spline2dSeg* Spline2d::mutable_smoothing_spline(const std::size_t index) {
+Spline2dSeg* Spline2d::mutable_smoothing_spline(const std::uint32_t index) {
   return &splines_[index];
 }
 
-const Spline2dSeg& Spline2d::smoothing_spline(const std::size_t index) const {
+const Spline2dSeg& Spline2d::smoothing_spline(const std::uint32_t index) const {
   return splines_[index];
 }
 
 const std::vector<double>& Spline2d::t_knots() const { return t_knots_; }
 
-std::size_t Spline2d::spline_order() const { return spline_order_; }
+std::uint32_t Spline2d::spline_order() const { return spline_order_; }
 
-std::size_t Spline2d::find_index(const double t) const {
+std::uint32_t Spline2d::find_index(const double t) const {
   auto upper_bound = std::upper_bound(t_knots_.begin() + 1, t_knots_.end(), t);
-  return std::min(t_knots_.size() - 1,
-                  static_cast<std::size_t>(upper_bound - t_knots_.begin())) - 1;
+  return std::min(static_cast<std::uint32_t>(t_knots_.size() - 1),
+                  static_cast<std::uint32_t>(upper_bound - t_knots_.begin())) -
+         1;
 }
 
 }  // namespace planning
